Feat & Fix: Changed timeout to retries
Given that Swift Actor integrated nicely with its async/await nicely, an
async function can await for a finished operation from the actors, thus
when expecting a returned value it can easily awaited.

The issue comes with actor isolation / mailbox system, it could possibly that the API
ask the inbox actor before the inbox was given value. In this scenario,
Actor will block the setter until the getter finishes like dispatch
queue.

So instead, the getter should be able to be called more than 1 time to
which increase the chance of a setter being queued before a getter.